Refactor dependency package.

+++ project/dependency/vendor.go
@@ -1,6 +1,7 @@
 package dependency
 
 import (
+	"errors"
 	"fmt"
 	"os"
 	"path/filepath"
@@ -10,21 +11,24 @@ import (
 )
 
 func (d *Dependency) Vendor(vendorRoot string) error {
+	if d.State == nil {
+		return errors.New("State has not been locked.")
+	}
 	var (
 		err  error
-		path = filepath.Clean(fmt.Sprintf("%s/%s/", vendorRoot, d.RootPackage))
+		path = filepath.Clean(fmt.Sprintf("%s/%s/", vendorRoot, d.rootPackage))
 	)
 
 	if err = os.RemoveAll(path); err != nil && !os.IsNotExist(err) {
 		return err
-	} else if err = os.Mkdir(path, 0775); err != nil {
+	} else if err = os.MkdirAll(path, 0775); err != nil {
 		return err
 	}
-	v := modules.GetVCS(d.VCS)
+	v := modules.GetVCS(d.vcs)
 	if v == nil {
-		return fmt.Errorf("Unkown Version Control System : %s", d.VCS)
+		return fmt.Errorf("Unkown Version Control System : %s", d.vcs)
 	}
-	d.Repository, err = vcs.CloneAtRevision(v, d.Remote, d.Revision, path)
+	d.repository, err = vcs.CloneAtRevision(v, d.remote, d.revision, path)
 	return err
 }
 
@@ -32,14 +36,14 @@ func (d *Dependency) Update(vendorRoot string) (updated bool, err error) {
 	var revision string
 	if err = d.Vendor(vendorRoot); err != nil {
 		return
-	} else if err = d.Repository.CheckoutBranch(d.Branch); err != nil {
+	} else if err = d.repository.CheckoutBranch(d.branch); err != nil {
 		return
-	} else if revision, err = d.Repository.GetRevision(); err != nil {
+	} else if revision, err = d.repository.GetRevision(); err != nil {
 		return
 	}
-	if d.Revision != revision {
+	if d.revision != revision {
 		updated = true
-		d.Revision = revision
+		d.revision = revision
 	}
 	return
 }
